Create Jason webshare welcome page
I created this page in Google sites, so I don't know all the tags it uses. I am using this as an arrival destination for my learning path. I know I don't need all the tags in this HTML file, but for now I will leave them in there, and "prune" them down in future versions.